# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest API - IPK",
        "description": "Dokumentasi penggunaan Rest API untuk berbagi pakai data Indeks Pembangunan Kebudayaan (IPK)",
        "contact": {
            "email": "pusdatin@kemenbud.go.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server API"
        }
    ],
    "paths": {
...
        "/api/data/dimensi-provinsi": {
            "get": {
                "tags": [
                    "3. Data"
                ],
                "summary": "Data dimensi provinsi",
                "description": "1. Mengembalikan semua data dimensi provinsi",
                "operationId": "getDimensiProvinsi",
                "parameters": [
                    {
                        "name": "kode",
                        "in": "query",
                        "description": "Kode provinsi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/data/indikator-provinsi": {
            "get": {
                "tags": [
                    "3. Data"
                ],
                "summary": "Data indikator provinsi",
                "description": "1. Mengembalikan semua data indikator provinsi",
                "operationId": "getIndikatorProvinsi",
                "parameters": [
                    {
                        "name": "kode",
                        "in": "query",
                        "description": "Kode provinsi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/api/referensi/indikator": {
            "get": {
                "tags": [
                    "1. Referensi"
                ],
                "summary": "Data referensi indikator",
                "description": "4. Mengembalikan semua data referensi indikator",
                "operationId": "getReferensiIndikatorList",
                "parameters": [
                    {
                        "name": "tahun",
                        "in": "query",
                        "description": "Tahun data",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    }
}